# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MAF Loyalty Engine API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CustomerTier = (typeof CustomerTier)[keyof typeof CustomerTier];

export const CustomerTier = {
  silver: "silver",
  gold: "gold",
  platinum: "platinum",
  diamond: "diamond",
} as const;

export interface Customer {
  id: number;
  name: string;
  email: string;
  phone?: string;
  tier: CustomerTier;
  totalPoints: number;
  joinedAt: string;
}

export interface Transaction {
  id: number;
  customerId: number;
  brandId: number;
  brandName?: string;
  amount: number;
  pointsEarned: number;
  description?: string;
  createdAt: string;
}

export type OfferStatus = (typeof OfferStatus)[keyof typeof OfferStatus];

export const OfferStatus = {
  active: "active",
  expired: "expired",
  redeemed: "redeemed",
} as const;

export interface Offer {
  id: number;
  customerId: number;
  brandId: number;
  brandName?: string;
  title: string;
  description: string;
  discountPercent: number;
  pointsBonus?: number;
  validUntil: string;
  status: OfferStatus;
  aiReason?: string;
  createdAt: string;
}

export interface BrandSpend {
  brandId: number;
  brandName: string;
  totalSpend: number;
  totalPoints: number;
  visitCount: number;
}

export type CustomerWithDetails = Customer & {
  transactions?: Transaction[];
  offers?: Offer[];
  brandSpend?: BrandSpend[];
};

export type CreateCustomerBodyTier =
  (typeof CreateCustomerBodyTier)[keyof typeof CreateCustomerBodyTier];

export const CreateCustomerBodyTier = {
  silver: "silver",
  gold: "gold",
  platinum: "platinum",
  diamond: "diamond",
} as const;

export interface CreateCustomerBody {
  name: string;
  email: string;
  phone?: string;
  tier?: CreateCustomerBodyTier;
}

export interface Brand {
  id: number;
  name: string;
  category: string;
  logoColor?: string;
  description?: string;
}

export interface CreateTransactionBody {
  customerId: number;
  brandId: number;
  amount: number;
  description?: string;
}

export interface GenerateOffersBody {
  customerId: number;
}

export interface GenerateOffersResponse {
  offers: Offer[];
  summary: string;
}

export type AnalyticsResponseTopBrandsItem = {
  brandName: string;
  revenue: number;
  customerCount: number;
};

export type AnalyticsResponseTierDistribution = {
  silver: number;
  gold: number;
  platinum: number;
  diamond: number;
};

export type AnalyticsResponseWeeklyOfferStats = {
  generated: number;
  redeemed: number;
  redemptionRate: number;
};

export interface AnalyticsResponse {
  totalCustomers: number;
  totalTransactions: number;
  totalPointsIssued: number;
  topBrands: AnalyticsResponseTopBrandsItem[];
  tierDistribution: AnalyticsResponseTierDistribution;
  weeklyOfferStats: AnalyticsResponseWeeklyOfferStats;
}

export type ListTransactionsParams = {
  customerId?: number;
  brandId?: number;
};

export type ListOffersParams = {
  customerId?: number;
  status?: ListOffersStatus;
};

export type ListOffersStatus =
  (typeof ListOffersStatus)[keyof typeof ListOffersStatus];

export const ListOffersStatus = {
  active: "active",
  expired: "expired",
  redeemed: "redeemed",
} as const;