# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MAF Loyalty Engine API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all customers
 */
export const ListCustomersResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string().optional(),
  tier: zod.enum(["silver", "gold", "platinum", "diamond"]),
  totalPoints: zod.number(),
  joinedAt: zod.date(),
});
export const ListCustomersResponse = zod.array(ListCustomersResponseItem);

/**
 * @summary Create a new customer
 */
export const CreateCustomerBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  phone: zod.string().optional(),
  tier: zod.enum(["silver", "gold", "platinum", "diamond"]).optional(),
});

/**
 * @summary Get customer by ID
 */
export const GetCustomerParams = zod.object({
  id: zod.coerce.number(),
});

export const GetCustomerResponse = zod
  .object({
    id: zod.number(),
    name: zod.string(),
    email: zod.string(),
    phone: zod.string().optional(),
    tier: zod.enum(["silver", "gold", "platinum", "diamond"]),
    totalPoints: zod.number(),
    joinedAt: zod.date(),
  })
  .and(
    zod.object({
      transactions: zod
        .array(
          zod.object({
            id: zod.number(),
            customerId: zod.number(),
            brandId: zod.number(),
            brandName: zod.string().optional(),
            amount: zod.number(),
            pointsEarned: zod.number(),
            description: zod.string().optional(),
            createdAt: zod.date(),
          }),
        )
        .optional(),
      offers: zod
        .array(
          zod.object({
            id: zod.number(),
            customerId: zod.number(),
            brandId: zod.number(),
            brandName: zod.string().optional(),
            title: zod.string(),
            description: zod.string(),
            discountPercent: zod.number(),
            pointsBonus: zod.number().optional(),
            validUntil: zod.date(),
            status: zod.enum(["active", "expired", "redeemed"]),
            aiReason: zod.string().optional(),
            createdAt: zod.date(),
          }),
        )
        .optional(),
      brandSpend: zod
        .array(
          zod.object({
            brandId: zod.number(),
            brandName: zod.string(),
            totalSpend: zod.number(),
            totalPoints: zod.number(),
            visitCount: zod.number(),
          }),
        )
        .optional(),
    }),
  );

/**
 * @summary List all MAF brands
 */
export const ListBrandsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.string(),
  logoColor: zod.string().optional(),
  description: zod.string().optional(),
});
export const ListBrandsResponse = zod.array(ListBrandsResponseItem);

/**
 * @summary List transactions
 */
export const ListTransactionsQueryParams = zod.object({
  customerId: zod.coerce.number().optional(),
  brandId: zod.coerce.number().optional(),
});

export const ListTransactionsResponseItem = zod.object({
  id: zod.number(),
  customerId: zod.number(),
  brandId: zod.number(),
  brandName: zod.string().optional(),
  amount: zod.number(),
  pointsEarned: zod.number(),
  description: zod.string().optional(),
  createdAt: zod.date(),
});
export const ListTransactionsResponse = zod.array(ListTransactionsResponseItem);

/**
 * @summary Record a new transaction
 */
export const CreateTransactionBody = zod.object({
  customerId: zod.number(),
  brandId: zod.number(),
  amount: zod.number(),
  description: zod.string().optional(),
});

/**
 * @summary List personalized offers
 */
export const ListOffersQueryParams = zod.object({
  customerId: zod.coerce.number().optional(),
  status: zod.enum(["active", "expired", "redeemed"]).optional(),
});

export const ListOffersResponseItem = zod.object({
  id: zod.number(),
  customerId: zod.number(),
  brandId: zod.number(),
  brandName: zod.string().optional(),
  title: zod.string(),
  description: zod.string(),
  discountPercent: zod.number(),
  pointsBonus: zod.number().optional(),
  validUntil: zod.date(),
  status: zod.enum(["active", "expired", "redeemed"]),
  aiReason: zod.string().optional(),
  createdAt: zod.date(),
});
export const ListOffersResponse = zod.array(ListOffersResponseItem);

/**
 * @summary Redeem an offer
 */
export const RedeemOfferParams = zod.object({
  id: zod.coerce.number(),
});

export const RedeemOfferResponse = zod.object({
  id: zod.number(),
  customerId: zod.number(),
  brandId: zod.number(),
  brandName: zod.string().optional(),
  title: zod.string(),
  description: zod.string(),
  discountPercent: zod.number(),
  pointsBonus: zod.number().optional(),
  validUntil: zod.date(),
  status: zod.enum(["active", "expired", "redeemed"]),
  aiReason: zod.string().optional(),
  createdAt: zod.date(),
});

/**
 * @summary AI-generate personalized weekly offers for a customer
 */
export const GenerateOffersBody = zod.object({
  customerId: zod.number(),
});

export const GenerateOffersResponse = zod.object({
  offers: zod.array(
    zod.object({
      id: zod.number(),
      customerId: zod.number(),
      brandId: zod.number(),
      brandName: zod.string().optional(),
      title: zod.string(),
      description: zod.string(),
      discountPercent: zod.number(),
      pointsBonus: zod.number().optional(),
      validUntil: zod.date(),
      status: zod.enum(["active", "expired", "redeemed"]),
      aiReason: zod.string().optional(),
      createdAt: zod.date(),
    }),
  ),
  summary: zod.string(),
});

/**
 * @summary Get cross-brand analytics and insights
 */
export const GetAnalyticsResponse = zod.object({
  totalCustomers: zod.number(),
  totalTransactions: zod.number(),
  totalPointsIssued: zod.number(),
  topBrands: zod.array(
    zod.object({
      brandName: zod.string(),
      revenue: zod.number(),
      customerCount: zod.number(),
    }),
  ),
  tierDistribution: zod.object({
    silver: zod.number(),
    gold: zod.number(),
    platinum: zod.number(),
    diamond: zod.number(),
  }),
  weeklyOfferStats: zod.object({
    generated: zod.number(),
    redeemed: zod.number(),
    redemptionRate: zod.number(),
  }),
});