# lib/api-zod/src/generated/types/brandSpend.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MAF Loyalty Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface BrandSpend {
  brandId: number;
  brandName: string;
  totalSpend: number;
  totalPoints: number;
  visitCount: number;
}