# lib/api-zod/src/generated/types/createTransactionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MAF Loyalty Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTransactionBody {
  customerId: number;
  brandId: number;
  amount: number;
  description?: string;
}