# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MAF Loyalty Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerTier } from "./customerTier";

export interface Customer {
  id: number;
  name: string;
  email: string;
  phone?: string;
  tier: CustomerTier;
  totalPoints: number;
  joinedAt: Date;
}