# lib/api-zod/src/generated/types/generateOffersResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MAF Loyalty Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { Offer } from "./offer";

export interface GenerateOffersResponse {
  offers: Offer[];
  summary: string;
}