# lib/api-zod/src/generated/types/offer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MAF Loyalty Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { OfferStatus } from "./offerStatus";

export interface Offer {
  id: number;
  customerId: number;
  brandId: number;
  brandName?: string;
  title: string;
  description: string;
  discountPercent: number;
  pointsBonus?: number;
  validUntil: Date;
  status: OfferStatus;
  aiReason?: string;
  createdAt: Date;
}